add: error case tests
