+++ README.md
@@ -10,6 +10,9 @@ npm i releasify -g
 
 // npx usage
 npx releasify <command>
+
+// show man of a command
+npx releasify <command> --help
 ```
 
 ## Commands
@@ -19,12 +22,21 @@ to get a preview.
 
 ### Draft
 
-Print out the new version that should be released with its changelog based on the commit messages.
+Print out the new version that should be released with its changelog listing the commit messages.
+
+The commits shown are
++ from: the `HEAD` of your local project 
++ to: the _first commit_ or the _last version tag_ you created
+
+Moreover, if the commit message is written with the pattern: `text describing (#123)`, where
+`(#123)` is the pull request ID, the labels of that PR are downloaded and processed by the template
+engine. These info are fetched from GitHub, keep in mind that there are [rete limits](https://developer.github.com/v3/#rate-limiting).
 
 ```sh
 releasify draft [--path|-p <path>]
                 [--tag|-t <pattern>]
                 [--semver|-s <relese>]
+                [--verbose|-v <level>]
                 [--help|-h]
 ```
