+++ lib/args.js
@@ -16,7 +16,7 @@ module.exports = function parseArgs (args) {
     default: {
       help: false,
       path: process.cwd(),
-      verbose: 'error'
+      verbose: 'warn'
     }
   })
