+++ lib/commands/draft.js
@@ -15,7 +15,7 @@ const ARGS_SCHEMA = {
     tag: { type: 'string' },
     verbose: {
       type: 'string',
-      enum: [ 'debug', 'info', 'error' ]
+      enum: [ 'debug', 'info', 'warn', 'error' ]
     },
     semver: {
       type: 'string',
@@ -32,7 +32,6 @@ module.exports = async function (args) {
   logger.debug('Reading the project in %s', args.path)
   const git = GitDirectory(args.path)
 
-  // TODO manage errors to print better messages
   // TODO check git status!
 
   const pkg = git.getRepoPackage()
@@ -55,10 +54,15 @@ module.exports = async function (args) {
   const github = GitHub(pkg.repository.url)
   logger.info('Getting PR labels from GitHub..')
   for (const commit of commitsWithPR) {
-    const response = await github.getPRLabels(commit.pullRequest.id)
-    const labels = response.data.map(_ => _.name)
-    logger.debug('\tGet PR %s labels %o', commit.pullRequest.id, labels)
-    releaseBuilder.addLine(commit.message, labels)
+    try {
+      const response = await github.getPRLabels(commit.pullRequest.id)
+      const labels = response.data.map(_ => _.name)
+      logger.debug('\tGet PR %s labels %o', commit.pullRequest.id, labels)
+      releaseBuilder.addLine(commit.message, labels)
+    } catch (error) {
+      logger.warn(error, '\tError getting PR %s', commit.pullRequest.id)
+      releaseBuilder.addLine(commit.message, [])
+    }
   }
 
   // don't miss any commit
