+++ test/command-draft.test.js
@@ -1,99 +1,10 @@
 'use strict'
 
+const { join } = require('path')
 const t = require('tap')
-const proxyquire = require('proxyquire')
+const h = require('./helper')
 
-const cmd = proxyquire('../lib/commands/draft', {
-  '../git-directory': proxyquire('../lib/git-directory', {
-    'simple-git': function (path) {
-      return {
-        tag: function (commands, cb) {
-          const history = `dbcf234f456e00e262103c55cf2ce44f64fc1b7e
-55acd4af74a5e74cdf91f5cbc207c86a8b9fdcac
-797d63bbd491e20536d6cae8867c5210db75318e
-dcf63111cd7b062b1b935d9c4925f70c9e08fd05
-10cd7a46bd0ed5b6a4fbc54a7320c5bb76966c9e
-20e8b1bcbe6a28954a01a5a58de9cb9f9d616420
-1606f37011c58b556d07c2539203b95f895838f6
-946d9f84afb202e49c89d8511be14e9bf4740a1b
-b4560139a74d5a40c58d30d1bd7dd2dc21a1b153
-c037c5d6e3a601d0f2c66a31e09bb06fe7c123d2
-a9a2ba10618d9714d9d0d47d7aef4d779f57224f
-          `
-          cb(null, history)
-        },
-        log: function (params, cb) {
-          const commits = {
-            all: [ {
-              hash: '68241662e2328127cf571d46194b63da20ea55bc',
-              date: '2019-04-08 01:46:25 +0430',
-              message: 'TOC added to Reply.md (#1582)',
-              refs: 'HEAD -> master, origin/master, origin/HEAD',
-              body: '',
-              author_name: 'srmarjani',
-              author_email: 'razi.marjani@gmail.com' },
-            {
-              hash: 'b23a2a3923e6b644e910c89d4cabd1536c81922d',
-              date: '2019-04-07 20:10:01 +0800',
-              message: 'Update Routes.md (#1579)',
-              refs: '',
-              body: 'fix link',
-              author_name: 'Zoron',
-              author_email: 'zoronlivingston@gmail.com' },
-            {
-              hash: 'a41fb18899e41f40ca1d8e7cb9ace71fa04e373f',
-              date: '2019-04-06 19:27:39 +0200',
-              message: 'Update Ecosystem.md (#1570)',
-              refs: '',
-              body: '',
-              author_name: 'Giacomo Gregoletto',
-              author_email: 'greguz@users.noreply.github.com' },
-            {
-              hash: '60f51bf376a08bc728f1c8b21ac44446104f1d2a',
-              date: '2019-04-06 15:45:43 +0200',
-              message: 'doc add fastify-schema-constraint to ecosystem (#1573)',
-              refs: '',
-              body: '',
-              author_name: 'Manuel Spigolon',
-              author_email: 'behemoth89@gmail.com' },
-            {
-              hash: '8f2a65ae63670e3f5b2219361bc037f006e6c6eb',
-              date: '2019-04-06 15:43:29 +0200',
-              message: 'Fix ignoreTrailingSlash example (#1574)',
-              refs: '',
-              body: '',
-              author_name: 'Tommaso Allevi',
-              author_email: 'tomallevi@gmail.com' } ]
-          }
-
-          cb(null, commits)
-        }
-      }
-    }
-  }),
-  '../github': proxyquire('../lib/github', {
-    '@octokit/rest': function () {
-      return {
-        issues: {
-          listLabelsOnIssue: async function (options) {
-            return {
-              status: 200,
-              url: `https://api.github.com/repos/${options.owner}/${options.repo}/issues/${options.number}/labels`,
-              headers: {
-                server: 'Test.com',
-                status: '200OK'
-              },
-              data: [{ name: 'documentation' }],
-              pullRequest: {
-                id: options.number
-              }
-            }
-          }
-        }
-      }
-    }
-  })
-})
+const cmd = h.buildProxyCommand('../lib/commands/draft')
 
 const { test } = t
 const options = {
@@ -113,16 +24,54 @@ test('draft a version forced release', async t => {
   t.plan(3)
 
   const opts = Object.assign({}, options)
-  opts.path = require('path').join(__dirname, 'fake-project/')
+  opts.path = join(__dirname, 'fake-project/')
   opts.semver = 'major'
   delete opts.tag // autosense
 
   const build = await cmd(opts)
   t.equals(build.version, '12.0.0')
   t.equals(build.oldVersion, '11.14.42')
-  t.equals(build.message, '📚 PR:\n- Fix ignoreTrailingSlash example (#1574)\n- doc add fastify-schema-constraint to ecosystem (#1573)\n- Update Ecosystem.md (#1570)\n- Update Routes.md (#1579)\n- TOC added to Reply.md (#1582)\n')
+  t.equals(build.message, '📚 PR:\n- This is a commit without PR\n- doc add fastify-schema-constraint to ecosystem (#1573)\n- Update Ecosystem.md (#1570)\n- Update Routes.md (#1579)\n- TOC added to Reply.md (#1582)\n')
 })
 
-test('draft a suggested release') // TODO
+test('draft a suggested release', async t => {
+  t.plan(2)
 
-test('draft the first release') // TODO
+  const opts = Object.assign({}, options)
+  opts.path = join(__dirname, 'fake-project/')
+  delete opts.tag // autosense
+  delete opts.semver // auto-calculate
+
+  const build = await cmd(opts)
+  // TODO: now the suggestedRelease is not implemented
+  t.equals(build.version, '11.14.42')
+  t.equals(build.oldVersion, '11.14.42')
+})
+
+test('draft the first release', async t => {
+  t.plan(2)
+
+  const cmd = h.buildProxyCommand('../lib/commands/draft', { emptyTag: true })
+  const opts = Object.assign({}, options)
+  opts.path = join(__dirname, 'fake-project/')
+  opts.tag = 'bad-pattern'
+  delete opts.semver // auto-calculate
+
+  const build = await cmd(opts)
+  t.equals(build.version, '11.14.42')
+  t.equals(build.oldVersion, '11.14.42')
+})
+
+test('error management getting PR', async t => {
+  t.plan(2)
+
+  const cmd = h.buildProxyCommand('../lib/commands/draft', { githubThrow: true })
+  const opts = Object.assign({}, options)
+  opts.path = join(__dirname, 'fake-project/')
+  opts.tag = 'bad-pattern'
+  delete opts.semver // auto-calculate
+
+  const build = await cmd(opts)
+  t.equals(build.version, '11.14.42')
+  t.equals(build.oldVersion, '11.14.42')
+})
